neated uninitialized out detection code
